Keep the schema with the basemap which reads it
Nothing else asks a source what it calls things, so there was no one for
it to sit beside.
